+++ nikhilyates-two/src/app/page.tsx
@@ -1,7 +1,7 @@
 import LandingPage from "./landing/LandingPage";
 import Experience from "@/components/sections/experience/page";
 import About from "@/app/about/page";
-import { ContactForm } from "@/app/contact/page"
+import { ContactForm } from "@/app/contact/ContactForm"
 import Portfolio from "@/app/portfolio/page"
 
 export default function Home() {
